% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_xmls.R
\name{add_children_to_xml}
\alias{add_children_to_xml}
\title{Add new child nodes to xml}
\usage{
add_children_to_xml(
  filename,
  child_str,
  node_names,
  node_attrs,
  outfile,
  order = 0,
  create_parent = TRUE
)
}
\arguments{
\item{create_parent}{}
}
\value{

}
\description{
Add new child nodes to xml
}
