% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{find_matches}
\alias{find_matches}
\title{Find all matching nodes}
\usage{
find_matches(
  node,
  leaf_name,
  leaf_attrs = NULL,
  node_name = NULL,
  node_attrs = NULL
)
}
\arguments{
\item{node}{xml2 node object to search}

\item{leaf_name}{name of leaf}

\item{leaf_attrs}{named list of leaf attributes}

\item{node_name}{name of node}

\item{node_attrs}{named list of node attributes}
}
\value{
node list with all matching nodes
}
\description{
Find all matching nodes
}
