% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{get_file_stem}
\alias{get_file_stem}
\title{Extract root node only from xml file}
\usage{
get_file_stem(filename)
}
\arguments{
\item{filename}{name of the file}
}
\value{
xml2 object with only the root node information from the file
}
\description{
Extract root node only from xml file
}
