% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_xmls.R
\name{make_batch}
\alias{make_batch}
\title{Make a batch file for sets of input files}
\usage{
make_batch(
  input_dir,
  ensemble_set,
  outfile,
  outer_tag = "ComponentSet",
  inner_tag = "FileSet",
  sample_file = system.file("extdata", "batch_example.xml", package = "gcamxmlgen"),
  other_text = NULL
)
}
\arguments{
\item{input_dir}{folder where your input files are located}

\item{ensemble_set}{named list with names corresponding to folders containing sets of input files for GCAM}

\item{outfile}{name of batch file that will be created}

\item{outer_tag}{tag for outer set of nodes in batch file. Default is ComponentSet}

\item{inner_tag}{tag for inner set of nodes in batch file components. Default is FileSet}

\item{sample_file}{specific batch file to use as a base.}

\item{other_text}{named list with names matching names of ensemble set argument. This is used for any additional text that should be added after each file in that component set.}
}
\value{

}
\description{
Make a batch file for sets of input files
}
