% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_xmls.R
\name{set_value}
\alias{set_value}
\title{Set single node value}
\usage{
set_value(node, data, col_name)
}
\arguments{
\item{node}{xml2 node object}

\item{data}{Annual timeseries data in tabular format. year information should be in a column labeled "year"}

\item{col_name}{The name of the column with the data values}
}
\value{
this function modifies an existing node object and so does not have a return value
}
\description{
Set single node value
}
\details{
Set value of a single xml2 node object for a given year from tabular data
}
