% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_xmls.R
\name{set_xml}
\alias{set_xml}
\title{Set xml2 node values}
\usage{
set_xml(
  xml_tree,
  data,
  leaf_name,
  leaf_attrs = NULL,
  node_name = NULL,
  node_attrs = NULL,
  output_filename = NULL,
  col_name = ""
)
}
\arguments{
\item{xml_tree}{may be either xml file to be modified or xml2 tree object}

\item{data}{either a numerical value, or time series data as a tibble or dataframe. If data is tabular, it is assumed to be a time series and must have a column named 'year'}

\item{leaf_name}{xml name of the specific leaf node you want to change the value of}

\item{leaf_attrs}{named list with the name specifying the name of the attribute of the leaf you want to change. Value is the value of the attribute.}

\item{node_attrs}{list where the name is the xml node name and the value is the specific xml attribute 'name' within that node where the leaves you want to modify are found (later known as 'tags'). You can specify}

\item{output_filename}{file where modified xml should be saved}

\item{col_name}{if data is tabular, this is the column the values should be pulled from}
}
\value{
modified xml2 tree object
}
\description{
Set xml2 node values
}
\details{
Set multiple xml2 nodes to single value or time series data. This function returns the modified xml tree, and will also save an xml file of this tree if output_filename is specified.
By passing a tree instead of a file name, the user may use this function to iterate over multiple inputs to modify the same xml document before saving.
}
\examples{

}
