% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_utils.R
\name{strip_string}
\alias{strip_string}
\title{Remove odd characters in string}
\usage{
strip_string(string_in, sep = "_")
}
\arguments{
\item{string_in}{string to strip down by removing slashes, spaces, parentheses, and others}

\item{sep}{character to replace odd characters with}
}
\value{
simplified string without unwanted characters
}
\description{
Remove odd characters in string
}
\examples{
s <- "AIM/CGE 2.0_ADVANCE_2020_1.5C-2100"
strip_string(s)
}
